refactor : refactor part ui code
