+++ app/src/main/java/com/kalex/sp_aplication/presentation/composables/loadingIndicator.kt
@@ -1,0 +1,22 @@
+package com.kalex.sp_aplication.presentation.composables
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SophosLoadingIndicator() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ) {
+        CircularProgressIndicator(
+            modifier = Modifier
+                .fillMaxSize(0.1f),
+
+        )
+    }
+}
